docker container names and worker queue config using env vars  & factored out child process logic
