+++ worker/docker-compose.yml
@@ -5,10 +5,11 @@ services:
     build:
       context: .
       dockerfile: Dockerfile
-    ports:
-      - 3000:3000
+
     networks:
       - dredd-network
+    environment:
+      - QUEUE_NAME=${QUEUE_NAME}
 
 networks:
   dredd-network:
